requirements.txt
flask
gunicorn
torch
torchvision
timm
Pillow
jinja2
werkzeug
markupsafe